requirements.txt
asgiref==3.3.1
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.10.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.5
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
wincertstore==0.2
